+++ src/components/feature/amaranth/employee/EmpInfoBox.js
@@ -12,46 +12,14 @@ import EmpEmailSalarySelectBox from './EmpEmailSalarySelectBox';
 import { ko } from 'date-fns/esm/locale';
 
 const EmpInfoBox = ({
-  data,
   register,
-  openDate,
-  selectedValue,
-  handleRadioChange,
-  onChangeOpenPost,
-  address,
-  addressDetail,
-  setImage,
-  imgFile,
-  companyList,
   errors,
-  clickYN,
-  onFocusError,
-  errorName,
-  imgPriviewFile,
-  setImgPriviewFile,
-  handleOpenDateChange,
-  setCompany,
-  company,
-  workplaceList,
-  setWorkplaceList,
-  workplaceSelect,
-  setWorkplaceSelect,
-  onChangeTel,
-  onChangeHomeTel,
-  infoBoxEnrlData,
-  setInfoBoxEnrlData,
-  setChangeFormData,
-  onChangePersonalMAIL,
-  onChangeSalaryMAIL,
-  setEmailPersonalData,
-  emailPersonalData,
-  emailSalaryData,
-  setEmailSalaryData,
-  onChangeDBDataSearch,
-  getValues,
-  setError,
+  state,
+  actions,
+  setValue,
   clearErrors,
-  checkDBErrorYN,
+  setError,
+  getValues,
 }) => {
   const imgRef = useRef();
 
@@ -61,7 +29,7 @@ const EmpInfoBox = ({
     reader.readAsDataURL(fileBlob);
     return new Promise(resolve => {
       reader.onload = () => {
-        setImgPriviewFile(reader.result);
+        state.setImgPriviewFile(reader.result);
         console.log('base64:', reader.result);
         resolve();
       };
@@ -84,19 +52,19 @@ const EmpInfoBox = ({
                     type="file"
                     accept="image/*"
                     onChange={e => {
-                      setImage(e.target.files[0]);
+                      state.setImage(e.target.files[0]);
                       imgPreview(e.target.files[0]);
                     }}
                     ref={imgRef}
                   />
                   <div className="userImageLabel">
-                    {imgPriviewFile ? (
-                      <img src={imgPriviewFile} alt="userImage" />
+                    {state.imgPriviewFile ? (
+                      <img src={state.imgPriviewFile} alt="userImage" />
                     ) : (
                       <img
                         src={
-                          imgFile
-                            ? imgFile
+                          state.imgFile
+                            ? state.imgFile
                             : 'https://upload.wikimedia.org/wikipedia/commons/thumb/5/59/User-avatar.svg/2048px-User-avatar.svg.png'
                         }
                         alt="userImage"
@@ -122,15 +90,15 @@ const EmpInfoBox = ({
                 <div className="empInfoCompanySelectBox">
                   <EmpInfoCompanySelectBox
                     width={230}
-                    data={companyList}
-                    setWorkplaceSelect={setWorkplaceSelect}
-                    setCompany={setCompany}
-                    company={company}
-                    setWorkplaceList={setWorkplaceList}
-                    clickYN={clickYN}
+                    data={state.companyList}
+                    setWorkplaceSelect={state.setWorkplaceSelect}
+                    setCompany={state.setCompany}
+                    company={state.company}
+                    setWorkplaceList={state.setWorkplaceList}
+                    clickYN={state.clickYN}
                     register={register}
                     errors={errors}
-                    errorName={errorName}
+                    errorName={state.errorName}
                     setError={setError}
                     getValues={getValues}
                     clearErrors={clearErrors}
@@ -139,14 +107,14 @@ const EmpInfoBox = ({
                 <div className="empInfoWorkplaceSelectBox">
                   <EmpInfoWorkplaceSelectBox
                     width={'calc(100% - 7px)'}
-                    data={workplaceList || []}
-                    workplaceSelect={workplaceSelect}
-                    setWorkplaceSelect={setWorkplaceSelect}
-                    clickYN={clickYN}
+                    data={state.workplaceList || []}
+                    workplaceSelect={state.workplaceSelect}
+                    setWorkplaceSelect={state.setWorkplaceSelect}
+                    clickYN={state.clickYN}
                     register={register}
                     errors={errors}
-                    errorName={errorName}
-                    setChangeFormData={setChangeFormData}
+                    errorName={state.errorName}
+                    setChangeFormData={state.setChangeFormData}
                   />
                 </div>
               </div>
@@ -163,25 +131,27 @@ const EmpInfoBox = ({
                   placeholder="사원번호를 입력해주세요(최대 10글자)"
                   {...register(
                     'emp_CD',
-                    !clickYN && {
+                    !state.clickYN && {
                       required: '사번을 입력해주세요.',
                     }
                   )}
                   maxLength="10"
-                  defaultValue={data?.emp_CD}
+                  defaultValue={state.data?.emp_CD}
                   style={{
                     border:
                       errors?.emp_CD &&
-                      (checkDBErrorYN?.emp_CD_ERROR || errorName === 'emp_CD')
+                      (state.checkDBErrorYN?.emp_CD_ERROR ||
+                        state.errorName === 'emp_CD')
                         ? '1px solid red'
                         : '1px solid #ccc',
-                    backgroundColor: clickYN ? '#f2f2f2' : '#fef4f4',
+                    backgroundColor: state.clickYN ? '#f2f2f2' : '#fef4f4',
                   }}
-                  disabled={clickYN}
-                  onFocus={onFocusError}
-                  onChange={onChangeDBDataSearch}
+                  disabled={state.clickYN}
+                  onFocus={actions.onFocusError}
+                  onChange={actions.onChangeDBDataSearch}
                 />
-                {(checkDBErrorYN.emp_CD_ERROR || errorName === 'emp_CD') && (
+                {(state.checkDBErrorYN.emp_CD_ERROR ||
+                  state.errorName === 'emp_CD') && (
                   <ErrorMessage
                     errors={errors}
                     name="emp_CD"
@@ -203,18 +173,18 @@ const EmpInfoBox = ({
                   placeholder="사용자 이름을 입력해주세요"
                   {...register(
                     'kor_NM',
-                    !clickYN && { required: '이름을 입력해주세요.' }
+                    !state.clickYN && { required: '이름을 입력해주세요.' }
                   )}
-                  defaultValue={data.kor_NM}
+                  defaultValue={state.data.kor_NM}
                   style={{
                     border:
-                      errors.kor_NM && errorName === 'kor_NM'
+                      errors.kor_NM && state.errorName === 'kor_NM'
                         ? '1px solid red'
                         : '1px solid #ccc',
                   }}
-                  onFocus={onFocusError}
+                  onFocus={actions.onFocusError}
                 />
-                {errorName === 'kor_NM' && (
+                {state.errorName === 'kor_NM' && (
                   <ErrorMessage
                     errors={errors}
                     name="kor_NM"
@@ -236,25 +206,25 @@ const EmpInfoBox = ({
                   name="username"
                   {...register(
                     'username',
-                    !clickYN && { required: 'ID를 입력해주세요.' }
+                    !state.clickYN && { required: 'ID를 입력해주세요.' }
                   )}
-                  defaultValue={data?.username}
+                  defaultValue={state.data?.username}
                   maxLength="16"
                   style={{
                     border:
                       errors?.username &&
-                      (checkDBErrorYN?.username_ERROR ||
-                        errorName === 'username')
+                      (state.checkDBErrorYN?.username_ERROR ||
+                        state.errorName === 'username')
                         ? '1px solid red'
                         : '1px solid #ccc',
-                    backgroundColor: clickYN ? '#f2f2f2' : '#fef4f4',
+                    backgroundColor: state.clickYN ? '#f2f2f2' : '#fef4f4',
                   }}
-                  disabled={clickYN}
-                  onFocus={onFocusError}
-                  onChange={onChangeDBDataSearch}
+                  disabled={state.clickYN}
+                  onFocus={actions.onFocusError}
+                  onChange={actions.onChangeDBDataSearch}
                 />
-                {(checkDBErrorYN?.username_ERROR ||
-                  errorName === 'username') && (
+                {(state.checkDBErrorYN?.username_ERROR ||
+                  state.errorName === 'username') && (
                   <ErrorMessage
                     errors={errors}
                     name="username"
@@ -273,7 +243,7 @@ const EmpInfoBox = ({
                   className="reqInputStyle"
                   {...register(
                     'email_ADD',
-                    !clickYN && {
+                    !state.clickYN && {
                       required: 'ID를 입력해주세요.',
                       pattern: {
                         value: /\w+/,
@@ -281,22 +251,22 @@ const EmpInfoBox = ({
                       },
                     }
                   )}
-                  defaultValue={data.email_ADD}
+                  defaultValue={state.data.email_ADD}
                   style={{
                     border:
                       errors?.email_ADD &&
-                      (checkDBErrorYN?.email_ADD_ERROR ||
-                        errorName === 'email_ADD')
+                      (state.checkDBErrorYN?.email_ADD_ERROR ||
+                        state.errorName === 'email_ADD')
                         ? '1px solid red'
                         : '1px solid #ccc',
-                    backgroundColor: clickYN ? '#f2f2f2' : '#fef4f4',
+                    backgroundColor: state.clickYN ? '#f2f2f2' : '#fef4f4',
                   }}
-                  disabled={clickYN}
-                  onFocus={onFocusError}
-                  onChange={onChangeDBDataSearch}
+                  disabled={state.clickYN}
+                  onFocus={actions.onFocusError}
+                  onChange={actions.onChangeDBDataSearch}
                 />
-                {(checkDBErrorYN?.email_ADD_ERROR ||
-                  errorName === 'email_ADD') && (
+                {(state.checkDBErrorYN?.email_ADD_ERROR ||
+                  state.errorName === 'email_ADD') && (
                   <ErrorMessage
                     errors={errors}
                     name="email_ADD"
@@ -317,19 +287,19 @@ const EmpInfoBox = ({
                   className="reqInputStyle"
                   {...register(
                     'password',
-                    !clickYN && {
+                    !state.clickYN && {
                       required: '비밀번호를 입력해주세요.',
                     }
                   )}
                   style={{
                     border:
-                      errors.password && errorName === 'password'
+                      errors.password && state.errorName === 'password'
                         ? '1px solid red'
                         : '1px solid #ccc',
                   }}
-                  onFocus={onFocusError}
+                  onFocus={actions.onFocusError}
                 />
-                {errorName === 'password' && (
+                {state.errorName === 'password' && (
                   <ErrorMessage
                     errors={errors}
                     name="password"
@@ -353,8 +323,8 @@ const EmpInfoBox = ({
                   type="radio"
                   name="gender_FG"
                   value="W"
-                  checked={selectedValue === 'W'}
-                  onChange={handleRadioChange}
+                  checked={state.selectedValue === 'W'}
+                  onChange={actions.handleRadioChange}
                 />
                 여자
                 <input
@@ -362,8 +332,8 @@ const EmpInfoBox = ({
                   type="radio"
                   name="gender_FG"
                   value="M"
-                  checked={selectedValue === 'M'}
-                  onChange={handleRadioChange}
+                  checked={state.selectedValue === 'M'}
+                  onChange={actions.handleRadioChange}
                 />
                 남자
               </div>
@@ -373,13 +343,13 @@ const EmpInfoBox = ({
               <EmpInfoEnrlSelectBox
                 width={'calc(100% - 7px)'}
                 data={['재직', '휴직', '퇴직']}
-                infoBoxEnrlData={infoBoxEnrlData}
-                setInfoBoxEnrlData={setInfoBoxEnrlData}
-                clickYN={clickYN}
+                infoBoxEnrlData={state.infoBoxEnrlData}
+                setInfoBoxEnrlData={state.setInfoBoxEnrlData}
+                clickYN={state.clickYN}
                 register={register}
                 errors={errors}
-                errorName={errorName}
-                setChangeFormData={setChangeFormData}
+                errorName={state.errorName}
+                setChangeFormData={state.setChangeFormData}
               />
             </td>
           </tr>
@@ -392,7 +362,7 @@ const EmpInfoBox = ({
                   className="mailInputStyle"
                   name="personal_MAIL"
                   {...register('personal_MAIL')}
-                  defaultValue={data?.personal_MAIL}
+                  defaultValue={state.data?.personal_MAIL}
                   maxLength="32"
                 />
                 @
@@ -402,14 +372,14 @@ const EmpInfoBox = ({
                   name="personal_MAIL_CP"
                   maxLength="32"
                   {...register('personal_MAIL_CP')}
-                  defaultValue={data?.personal_MAIL_CP}
+                  defaultValue={state.data?.personal_MAIL_CP}
                 />
                 <EmpEmailPersonalSelectBox
                   width={'calc(100% - 300px)'}
-                  onClickEvent={onChangePersonalMAIL}
-                  emailPersonalData={emailPersonalData}
-                  setEmailPersonalData={setEmailPersonalData}
-                  setChangeFormData={setChangeFormData}
+                  onClickEvent={actions.onChangePersonalMAIL}
+                  emailPersonalData={state.emailPersonalData}
+                  setEmailPersonalData={state.setEmailPersonalData}
+                  setChangeFormData={state.setChangeFormData}
                 />
               </div>
             </td>
@@ -424,7 +394,7 @@ const EmpInfoBox = ({
                   name="salary_MAIL"
                   maxLength="32"
                   {...register('salary_MAIL')}
-                  defaultValue={data?.salary_MAIL}
+                  defaultValue={state.data?.salary_MAIL}
                 />
                 @
                 <input
@@ -433,14 +403,14 @@ const EmpInfoBox = ({
                   name="salary_MAIL_CP"
                   maxLength="32"
                   {...register('salary_MAIL_CP')}
-                  defaultValue={data?.salary_MAIL_CP}
+                  defaultValue={state.data?.salary_MAIL_CP}
                 />
                 <EmpEmailSalarySelectBox
                   width={'calc(100% - 300px)'}
-                  onClickEvent={onChangeSalaryMAIL}
-                  emailSalaryData={emailSalaryData}
-                  setEmailSalaryData={setEmailSalaryData}
-                  setChangeFormData={setChangeFormData}
+                  onClickEvent={actions.onChangeSalaryMAIL}
+                  emailSalaryData={state.emailSalaryData}
+                  setEmailSalaryData={state.setEmailSalaryData}
+                  setChangeFormData={state.setChangeFormData}
                 />
               </div>
             </td>
@@ -453,8 +423,8 @@ const EmpInfoBox = ({
                 className="inputStyle"
                 name="tel"
                 {...register('tel')}
-                onChange={onChangeTel}
-                defaultValue={data?.tel}
+                onChange={actions.onChangeTel}
+                defaultValue={state.data?.tel}
                 maxLength="13"
               />
             </td>
@@ -465,8 +435,8 @@ const EmpInfoBox = ({
                 name="home_TEL"
                 className="inputStyle"
                 {...register('home_TEL')}
-                onChange={onChangeHomeTel}
-                defaultValue={data?.home_TEL}
+                onChange={actions.onChangeHomeTel}
+                defaultValue={state.data?.home_TEL}
                 maxLength="13"
               />
             </td>
@@ -481,11 +451,13 @@ const EmpInfoBox = ({
                 className="addressInputStyle"
                 name="zipcode"
                 {...register('zipcode')}
-                defaultValue={address ? address : data?.zipcode}
+                defaultValue={
+                  state.address ? state.address : state.data?.zipcode
+                }
               />
               <EventButton
                 data={'우편번호'}
-                onClickEvent={onChangeOpenPost}
+                onClickEvent={actions.onChangeOpenPost}
               ></EventButton>
             </td>
           </tr>
@@ -496,7 +468,9 @@ const EmpInfoBox = ({
                 className="addrNum"
                 name="addr"
                 {...register('addr')}
-                defaultValue={addressDetail ? addressDetail : data?.addr}
+                defaultValue={
+                  state.addressDetail ? state.addressDetail : state.data?.addr
+                }
               />
             </td>
             <td className="cellStyle">
@@ -505,7 +479,7 @@ const EmpInfoBox = ({
                 className="inputStyle"
                 name="addr_NUM"
                 {...register('addr_NUM')}
-                defaultValue={data.addr_NUM}
+                defaultValue={state.data.addr_NUM}
               />
             </td>
           </tr>
@@ -513,9 +487,9 @@ const EmpInfoBox = ({
             <th className="headerCellStyle">입사일</th>
             <td className="cellStyle">
               <DatePicker
-                selected={openDate}
+                selected={state.openDate}
                 name="join_DT"
-                onChange={handleOpenDateChange}
+                onChange={actions.handleOpenDateChange}
                 dateFormat="yyyy-MM-dd"
                 className="datePickerInputStyle"
                 locale={ko}
